explain 'the security of your encryption'
No one ever goes into settings looking for an option to 'verify the security of his encryption'. That is just not something a user might be thinking. What a user might go look for though, is a way to verify that he is really chatting with the person he presumes to be chatting with. That is what this feature is for, and that is what this text should convey.