differentiate photo, picture, image and avatar
Where speaking of image we are referring to the captured or stored information which describes a picture. An image viewer renders you a picture of the image file. It doesn't show you the complete image, as the image includes metadata.
When you share an image in a conversation however, you intent is sharing the whole image file, not only the picture which results from it, which is why we use the word image when sharing an image.

The word "photo" should be reserved for places where it is highly likely that the image is actually a photo rather than any other kind of image.

The word avatar should not be used, as it is often understood as a fictional representation of a person, rather than a realistic representation. Avatars are used in games. Pictures and photos relate better to the real world.

For profiles, use "profile picture" rather than avatar, image or photo. We don't know if the user chooses to set a photo or a drawing of his face, so "picture" fits better than "photo".
For your profile you upload the whole image, but the intent is not to share it as a image file, but rather to render a visible picture within Signal. Contacts are not offered to save your profile picture as an image, it really is only meant to represent you within the app itself. For this case the word "picture" fits better than "image".