Add Orbot and socks proxy support
There is a new 'Network' menu item in the settings to configure the
proxy. Orbot can be specified as proxy server, or the SOCKS5 server
address can be entered manually. It is also possible to setup the proxy
prior to registration.

When a proxy is set all outbound connections are tunneled, including
Websocket and DNS lookups (with DOH). For calls, TCP-only and TURN
routing is enforced in WebRTC.

It uses NetCipher library to start Orbot automatically if needed.

Closes #12