Fix reproducible build issues with Git hashes and apkdiff
We stop using abbreviated Git hashes and use full hashes instead for the
debug log output. According to
https://github.com/git/git/commit/e6c587c733b4634030b353f4024794b08bc86892,
the default abbreviation is subject to auto sizing depending on the
estimated number of objects in the Git repository. This number could
differ if the Signal developers use private branches for developing but
then only push the squashed commits to the public repository, or if there
are other differences in how the Git repository was cloned or configured.
This means the abbreviated Git hash length might not be deterministic.

To solve this, we just use the full Git hash. The GIT_HASH config field
is only used in the debug log print out anyway, and using the full hash
instead of a hardcoded abbreviated size helps future-proof it as well.

For apkdiff, we ignore the new file names for signing info. We keep the
older versions in case they change again (or if someone wants to verify
a previous version of Signal that had the old names).  Also, the
resources.arsc issue seems to be resolved according to the Google Issue
and local testing.

Fixes #10476